:gear: component refactor, edit route with links
