+++ src/components/version_bar.js
@@ -1,0 +1,14 @@
+import React  from "react"
+
+const Version = "0.6.4"
+
+const Bar = () =>
+    <div className="construction" >
+        <div>
+            <div className="content">
+                <span><i className="icon-beaker" /> version { Version } - early release</span>
+            </div>
+        </div>
+    </div>
+
+export default Bar
