+++ src/pages/admin.js
@@ -4,11 +4,15 @@ import { navigate, Link } from "gatsby"
 import Login from "../components/login";
 import PrivateRoute from "../components/privateRoute"
 import { getUser, isLoggedIn, logout, getAuthHeader } from "../services/auth"
+import Review from "../components/admin/review"
+import Edit from "../components/admin/edit"
+import Bar from "../components/version_bar"
 
 const Unsecure = () => 
     (<div className="container m-3">
         Utilisateur deconnecté
     </div>);
+
 
 const Secure = () =>  {
     const usr = getUser();
@@ -24,37 +28,6 @@ const Secure = () =>  {
     </div>);
 }
 
-const Review = () => {
-    // Client-side Runtime Data Fetching
-    const [reviewList, setReviewList] = useState([]);
-    useEffect(() => {
-        fetch(`http://localhost:5005/c/review`, {
-            headers : {"Authorization" : getAuthHeader()}
-        })
-            .then(response => response.json()) // parse JSON from request
-            .then(resultData => {
-                console.log(resultData);
-                setReviewList(resultData);
-            }) // set data for the number of stars
-    }, [])
-    return (
-        <div className="container m-3">
-            <p>Review</p>
-            <table>
-                <tr>
-                    <th>Id</th>
-                    <th>Ecli</th>
-                </tr>
-                { reviewList.map((item) => (
-                    <tr key={ item.id }>
-                        <td>{ item.id }</td>
-                        <td>{ item.ecli }</td>
-                    </tr>
-                ))}
-            </table>
-        </div>
-    );
-}
 
 const Admin = () => (
   <div>
@@ -64,7 +37,10 @@ const Admin = () => (
             <ul className="navbar-nav mr-auto">
                 <li className="nav-item"><Link className="nav-link" to="/admin">Home</Link></li>
                 { isLoggedIn() ? (
-                    <li className="nav-item"><Link className="nav-link" to="/admin/review">Review</Link></li>
+                    <>
+                        <li className="nav-item"><Link className="nav-link" to="/admin/review">Review</Link></li>
+                        <li className="nav-item"><Link className="nav-link" to="/admin/edit">Edit</Link></li>
+                    </>
                     ) : (
                     <li className="nav-item"><Link className="nav-link" to="/admin/login">Login</Link></li>
                     )
@@ -81,9 +57,11 @@ const Admin = () => (
             </ul>
         </div>
     </nav>
+    <Bar />
     <Router>
         { isLoggedIn() ? (<Secure path="/admin" />) : ( <Unsecure path="/admin" /> )}
         <PrivateRoute path="/admin/review" component={Review} />
+        <PrivateRoute path="/admin/edit/:docid" component={Edit} />
         { isLoggedIn() ? null : (<Login path="/admin/login" />) }
     </Router>
   </div>
